+++ routes/htmlRoutes.js
@@ -1,5 +1,4 @@
 var db = require("../models");
-var googleCivic = require("../api/googleCivic.js");
 
 var express = require("express");
 var router = express.Router();
@@ -12,13 +11,6 @@ module.exports = function(app) {
         msg: "Welcome!",
         examples: dbExamples
       });
-    });
-  });
-
-  router.get("/google-civic", function(req, res) {
-    googleCivic.all(function() {
-
-      res.render("test");
     });
   });
