+++ server.js
@@ -18,6 +18,9 @@ app.use(bodyParser.urlencoded({ extended: false }));
 app.use(bodyParser.json());
 app.use(express.static(__dirname + "/public"));
 
+// var routes = require("./routes/controller.js");
+// app.use(routes);
+
 // Handlebars
 app.engine(
   "handlebars",
